ADD TEST [AMD] Update the schedule in StreamPipeline when an inst has LocalLoadTrait
This can happen for LocalLoad (when using the gluon frontend) and also
for LocalLoadPackedTransposed since it's created in AccelerateAMDMatmul.
The rest of the code at the creates a LocalStoreOp and a LocalLoadOp
that will be deadcoded since there will be no uses of if as ahe other
uses have already been replaced in replaceUsesAndPropagateType.